chore: move LogSender to agentsdk
